Fix casting enum to int

+++ march_hardware/src/ethercat/pdo_map.cpp
@@ -39,7 +39,7 @@ void PDOmap::addObject(IMCObjectName object_name)
 
   if (this->PDO_objects.count(object_name) != 0)
   {
-    ROS_WARN("IMC object %i is already added to PDO map", object_name);
+    ROS_WARN("IMC object %i is already added to PDO map", (int)object_name);
     return;
   }
